Fix collisions with existing user and group ids
Do not add groups or users if ther gid or uid already exists in the base
image. We also make useradd and the USER instruction depend on numeric
ids since we can't guarantee that a group or user with name 'app'
exists.
The USER instruction now also includes the group id to ensure that any
created files have the expected user and group ownership.