Add --rebuild flag to force rebuilding the image
Currently, `rails-new` must be released to use any newly released
versions of Ruby/Rails by default. To address this, I'd like to remove
the fallback versions of Ruby/Rails in the CLI arguments. However, this
will mean that the image name no longer contains the Rails version, and
that can lead to a stale Rails version when running `rails-new` if it
has been previously built.

This commit addresses that by providing a `--rebuild` option which
forces the container to be rebuilt and ensures the latest version of
Rails is installed and used.